requirements.txt
ldap3==2.9.1
Flask==2.2.5
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Werkzeug==2.2.3
python-dotenv==1.0.1
SQLAlchemy>=2.0.16
pandas==2.2.1
numpy==1.24.3
openpyxl==3.1.2
xlrd==2.0.1
Flask-Dance[sqla]==7.0.0
requests==2.31.0
msal==1.28.0
XlsxWriter==3.1.9
pyodbc==5.0.1
pymssql==2.2.10
alembic==1.12.1
Flask-Migrate==4.0.5
pytz==2024.1
gunicorn==21.2.0

